Server frontend (#8)
Had to adjust some things with the P/R. Addressed a bug in registration logic. Removed files that shouldn't be commited to source control.

TODO chore:

For the various return statements in the routes, instead of returning standard strings, we should return JSON, since the client might be expecting that.

* Did some refactor work to prepare network layer for server loop layer. Fixed a slew of imports

* Lots of code restructure. More message mappers. A delegator to delegate message handler logic. Message mapper for game map. Main server loop is also included.

* Initial commit of server source

* Initial frontend server code

* Front-end server fixes

* Start of Authorization

* Created user database

Added register and login functionality using a Sqlite database. Also added a JWT token when user is authorized

* Removes files that should be ignored. Script is adjusted to address some bugs around registration. Script migrates database on start. Hash function adjusted to return a hex string. Various return statements return with HTTP statuses to better reflect various errors

Co-authored-by: Cody Huzarski <chuzarski@gmail.com>
Co-authored-by: Enqirue <excastillo@oakland.edu>